refactor: move to utils
