+++ packages/client/src/registry/StreamStorageRegistry.ts
@@ -1,22 +1,21 @@
+import { Provider } from '@ethersproject/providers'
+import { StreamID, toStreamID } from '@streamr/protocol'
+import { EthereumAddress, Logger, collect, toEthereumAddress } from '@streamr/utils'
+import min from 'lodash/min'
+import { Lifecycle, delay, inject, scoped } from 'tsyringe'
+import { Authentication, AuthenticationInjectionToken } from '../Authentication'
+import { ConfigInjectionToken, StrictStreamrClientConfig } from '../Config'
+import { ContractFactory } from '../ContractFactory'
+import { getStreamRegistryChainProviders, getStreamRegistryOverrides } from '../Ethereum'
+import { Stream } from '../Stream'
+import { StreamFactory } from '../StreamFactory'
+import { StreamIDBuilder } from '../StreamIDBuilder'
 import type { StreamStorageRegistryV2 as StreamStorageRegistryContract } from '../ethereumArtifacts/StreamStorageRegistryV2'
 import StreamStorageRegistryArtifact from '../ethereumArtifacts/StreamStorageRegistryV2Abi.json'
-import { scoped, Lifecycle, inject, delay } from 'tsyringe'
-import { Provider } from '@ethersproject/providers'
-import { ConfigInjectionToken, StrictStreamrClientConfig } from '../Config'
-import { Stream } from '../Stream'
-import { getStreamRegistryChainProviders, getStreamRegistryOverrides } from '../Ethereum'
-import { StreamID, toStreamID } from '@streamr/protocol'
-import { StreamIDBuilder } from '../StreamIDBuilder'
-import { waitForTx, queryAllReadonlyContracts, initContractEventGateway } from '../utils/contract'
-import { SynchronizedGraphQLClient } from '../utils/SynchronizedGraphQLClient'
 import { StreamrClientEventEmitter } from '../events'
-import { Authentication, AuthenticationInjectionToken } from '../Authentication'
-import { ContractFactory } from '../ContractFactory'
-import { EthereumAddress, Logger, toEthereumAddress } from '@streamr/utils'
 import { LoggerFactory } from '../utils/LoggerFactory'
-import { StreamFactory } from '../StreamFactory'
-import { collect } from '../utils/iterators'
-import min from 'lodash/min'
+import { SynchronizedGraphQLClient } from '../utils/SynchronizedGraphQLClient'
+import { initContractEventGateway, queryAllReadonlyContracts, waitForTx } from '../utils/contract'
 
 export interface StorageNodeAssignmentEvent {
     readonly streamId: StreamID
