+++ packages/client/test/integration/Sequencing.test.ts
@@ -1,14 +1,12 @@
 import 'reflect-metadata'
 
-import { wait, waitForCondition } from '@streamr/utils'
+import { collect, merge, wait, waitForCondition } from '@streamr/utils'
+import { Message } from '../../src/Message'
 import { Stream } from '../../src/Stream'
 import { StreamrClient } from '../../src/StreamrClient'
-import { collect } from '../../src/utils/iterators'
 import { FakeEnvironment } from '../test-utils/fake/FakeEnvironment'
 import { getWaitForStorage } from '../test-utils/publish'
 import { createTestStream, uid } from '../test-utils/utils'
-import { Message } from '../../src/Message'
-import { merge } from '@streamr/utils'
 
 const Msg = (opts?: any) => {
     return merge(
