+++ packages/client/test/integration/parallel-key-exchange.test.ts
@@ -1,17 +1,16 @@
 import 'reflect-metadata'
 
 import { Wallet } from '@ethersproject/wallet'
-import { wait, toEthereumAddress } from '@streamr/utils'
-import range from 'lodash/range'
 import { StreamMessageType } from '@streamr/protocol'
 import { fastWallet } from '@streamr/test-utils'
+import { collect, toEthereumAddress, wait } from '@streamr/utils'
+import range from 'lodash/range'
 import { createPrivateKeyAuthentication } from '../../src/Authentication'
+import { Stream } from '../../src/Stream'
+import { StreamrClient } from '../../src/StreamrClient'
 import { GroupKey } from '../../src/encryption/GroupKey'
 import { StreamPermission } from '../../src/permission'
 import { MessageFactory } from '../../src/publish/MessageFactory'
-import { Stream } from '../../src/Stream'
-import { StreamrClient } from '../../src/StreamrClient'
-import { collect } from '../../src/utils/iterators'
 import { createGroupKeyQueue, createStreamRegistryCached } from '../test-utils/utils'
 import { FakeEnvironment } from './../test-utils/fake/FakeEnvironment'
