+++ packages/client/test/integration/publisher-key-reuse.test.ts
@@ -2,10 +2,10 @@ import 'reflect-metadata'
 
 import { Wallet } from '@ethersproject/wallet'
 import { fastWallet } from '@streamr/test-utils'
-import { StreamPermission } from '../../src/permission'
+import { collect } from '@streamr/utils'
 import { Stream } from '../../src/Stream'
 import { StreamrClient } from '../../src/StreamrClient'
-import { collect } from '../../src/utils/iterators'
+import { StreamPermission } from '../../src/permission'
 import { FakeEnvironment } from '../test-utils/fake/FakeEnvironment'
 import { nextValue } from './../../src/utils/iterators'
