+++ packages/client/test/unit/searchStreams.test.ts
@@ -3,9 +3,9 @@ import 'reflect-metadata'
 import { BigNumber } from '@ethersproject/bignumber'
 import { StreamID, toStreamID } from '@streamr/protocol'
 import { randomEthereumAddress } from '@streamr/test-utils'
-import { searchStreams, SearchStreamsResultItem } from '../../src/registry/searchStreams'
+import { collect } from '@streamr/utils'
 import { Stream } from '../../src/Stream'
-import { collect } from '../../src/utils/iterators'
+import { SearchStreamsResultItem, searchStreams } from '../../src/registry/searchStreams'
 import { SynchronizedGraphQLClient } from '../../src/utils/SynchronizedGraphQLClient'
 import { mockLoggerFactory } from '../test-utils/utils'
