+++ packages/client/test/unit/subscribePipeline.test.ts
@@ -1,25 +1,24 @@
 import 'reflect-metadata'
 
 import { Wallet } from '@ethersproject/wallet'
-import { toEthereumAddress } from '@streamr/utils'
 import { EncryptionType, MessageID, StreamMessage, StreamPartID, StreamPartIDUtils, toStreamID } from '@streamr/protocol'
 import { fastWallet, randomEthereumAddress } from '@streamr/test-utils'
-import { Stream } from '../../src/Stream'
-import { createPrivateKeyAuthentication } from '../../src/Authentication'
-import { DestroySignal } from '../../src/DestroySignal'
-import { DecryptError, EncryptionUtil } from '../../src/encryption/EncryptionUtil'
-import { createSignedMessage } from '../../src/publish/MessageFactory'
-import { createSubscribePipeline } from "../../src/subscribe/subscribePipeline"
-import { collect } from '../../src/utils/iterators'
-import { mockLoggerFactory } from '../test-utils/utils'
-import { GroupKey } from './../../src/encryption/GroupKey'
-import { MessageStream } from './../../src/subscribe/MessageStream'
+import { collect, toEthereumAddress } from '@streamr/utils'
 import { mock } from 'jest-mock-extended'
+import { createPrivateKeyAuthentication } from '../../src/Authentication'
+import { StrictStreamrClientConfig } from '../../src/Config'
+import { DestroySignal } from '../../src/DestroySignal'
+import { Stream } from '../../src/Stream'
+import { DecryptError, EncryptionUtil } from '../../src/encryption/EncryptionUtil'
 import { GroupKeyManager } from '../../src/encryption/GroupKeyManager'
 import { LitProtocolFacade } from '../../src/encryption/LitProtocolFacade'
 import { SubscriberKeyExchange } from '../../src/encryption/SubscriberKeyExchange'
 import { StreamrClientEventEmitter } from '../../src/events'
-import { StrictStreamrClientConfig } from '../../src/Config'
+import { createSignedMessage } from '../../src/publish/MessageFactory'
+import { createSubscribePipeline } from "../../src/subscribe/subscribePipeline"
+import { mockLoggerFactory } from '../test-utils/utils'
+import { GroupKey } from './../../src/encryption/GroupKey'
+import { MessageStream } from './../../src/subscribe/MessageStream'
 
 const CONTENT = {
     foo: 'bar'
